+++ packages/dashboard/src/components/app.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import 'typeface-roboto';
 import { BASE_PATH, LOGIN_ROUTE } from '../util/url';
-import { AppContextProvider } from './app-context-provider';
+import { AppContextProvider } from './app-contexts';
 import './app.css';
 import Login from './auth/login';
 import PrivateRoute from './auth/private-route';
@@ -13,11 +13,7 @@ import { RmfContextProvider } from './rmf-contexts';
 export default function App(): JSX.Element {
   return (
     <AppContextProvider>
-      <RmfContextProvider
-        doorStates={doorStates}
-        liftStates={liftStates}
-        dispenserStates={dispenserStates}
-      >
+      <RmfContextProvider>
         <BrowserRouter>
           <Switch>
             <Route exact={true} path={LOGIN_ROUTE}>
